KNN/pyproject.toml
[project]
name = "knn"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy (>=2.3.2,<3.0.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]

[tool.poetry]
packages = [{include = "knn", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pylint = "^3.3.8"
pytest-cov = "^6.3.0"
coverage = "^7.10.6"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
